requirements.txt
flask==3.1.2
streamlit==1.49.1
python-dotenv==1.1.1
openai==1.102.0
langchain==0.3.27
langchain-openai==0.3.32
langchain-community==0.3.29
langsmith==0.4.21
ddgs==9.5.4